+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Add `ParegRef`. It allows parsing to continue even if there are references to
   the original arguments.
 - Add checkers: `check::CheckRef`, `check::InRangeI` and `check::InRange`.
+- Add new parsers `split_arg` and `arg_list` and their respective methods on
+  `ParegRef` and `Pareg`.
 
 ### Changes
 - Rename `Pareg::map_err` to `Pareg::map_res`, add `Pareg::map_err`.